batch: Extract replacement line runs
Ownership translation accepts replacement run context when live-hunk
selections need full-file before/after boundaries.

The run data and derivation sit in batch.ownership even though they
describe semantic line correspondence rather than ownership metadata.

This commit moves ReplacementLineRun and its derivation into
batch.replacement_line_runs and updates production callers to import that
owner.

Subsequent commits will move tests and architecture docs to the
replacement-run boundary.
